Fix skills status effects and summons.
With `filter_map`, we wouldn't `break` on time as we used to. While
parsing status effects, we wouldn't stop at the first `h4` we encounter.
`descend_if_tag` returns `None`, because the tag doesn't match, but it
does not stop the `filter_map`. `map_while` is stopped upon hitting an
`h4` now.