Add `json fetch_all_matches_from_guide`.
This ugly-looking command has a very wrong name, but I don't know how to
name it otherwise.
It looks for all codex URIs we have for all entities on the guide. It
loops over them and tries to fetch them all on the codex.

I used this for "importing" entries that I missed from previous events
and that were missing from backups.